refactor : apilog 패키지 분리
