+++ data/src/main/kotlin/com/wepli/data/db/WePLiDatabase.kt
@@ -6,8 +6,8 @@ import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverter
 import androidx.room.TypeConverters
-import com.wepli.data.devmode.dao.ApiLogDao
-import com.wepli.data.devmode.entity.ApiLogEntity
+import com.wepli.data.devmode.apilog.dao.ApiLogDao
+import com.wepli.data.devmode.apilog.entity.ApiLogEntity
 import kotlinx.serialization.json.Json
 
 // TypeConverter 클래스 생성
