+++ data/src/main/kotlin/com/wepli/data/devmode/apilog/datasource/DebugApiLogLocalDatasourceImpl.kt
@@ -1,7 +1,7 @@
-package com.wepli.data.devmode.datasource
+package com.wepli.data.devmode.apilog.datasource
 
-import com.wepli.data.devmode.dao.ApiLogDao
-import com.wepli.data.devmode.entity.ApiLogEntity
+import com.wepli.data.devmode.apilog.dao.ApiLogDao
+import com.wepli.data.devmode.apilog.entity.ApiLogEntity
 import javax.inject.Inject
 
 class DebugApiLogLocalDatasourceImpl @Inject constructor(
