+++ data/src/main/kotlin/com/wepli/data/devmode/apilog/repository/DebugApiLogRepositoryImpl.kt
@@ -1,8 +1,8 @@
-package com.wepli.data.devmode.repository
+package com.wepli.data.devmode.apilog.repository
 
-import com.wepli.data.devmode.datasource.DebugApiLogLocalDatasourceImpl
-import com.wepli.data.devmode.entity.toDomain
-import com.wepli.data.devmode.entity.toEntity
+import com.wepli.data.devmode.apilog.datasource.DebugApiLogLocalDatasourceImpl
+import com.wepli.data.devmode.apilog.entity.toDomain
+import com.wepli.data.devmode.apilog.entity.toEntity
 import com.wepli.domain.devmode.model.ApiLog
 import com.wepli.domain.devmode.repository.DebugApiLogRepository
 import javax.inject.Inject
