+++ data/src/main/kotlin/com/wepli/data/di/datasource/DataSourceModule.kt
@@ -9,8 +9,8 @@ import com.wepli.data.chart.datasource.remote.ChartDataSourceImpl
 import com.wepli.data.datastore.local.DataStorePrefDataSource
 import com.wepli.data.datastore.local.DataStorePrefDataSourceImpl
 import com.wepli.data.di.qualifier.RemoteDataSource
-import com.wepli.data.devmode.datasource.DebugApiLogLocalDatasource
-import com.wepli.data.devmode.datasource.DebugApiLogLocalDatasourceImpl
+import com.wepli.data.devmode.apilog.datasource.DebugApiLogLocalDatasource
+import com.wepli.data.devmode.apilog.datasource.DebugApiLogLocalDatasourceImpl
 import com.wepli.data.playlist.datasource.remote.PlaylistDataSource
 import com.wepli.data.playlist.datasource.remote.PlaylistDataSourceImpl
 import com.wepli.data.relaylist.datasource.remote.RelaylistDataSource
